refactor(api): :wrench: rename user views for consistency and clarity…

+++ core/store/api/v1/cart/services/delete_cart_product.py
@@ -7,5 +7,9 @@ class DeleteCartProductService:
 
     @staticmethod
     @transaction.atomic
-    def execute(cart_product: CartProduct) -> None:
+    def execute(cart_product_id : int) -> None:
+        cart_product = CartProduct.objects.get(id=cart_product_id)
+        if not cart_product:
+            raise CartProduct.DoesNotExist("Cart product not found")
+
         cart_product.delete()
